Complete functionalities and display of content
I pretty much finished the project by building all
necessary functions and finishing up on the display
of all weather data which I want to be displayed in
my app.

Searching for a location can sometimes end in an
error since the API expects the name of a city to
be strictly in English. Also, searching for a location
which exists multiple times will display one quite
random choice of possible locations with a shared name.

It would be great to add some auto-complete functionality
which automatically detects the possible options and
offers them as a choice while formatting the search
query to match the format required by the API.